demo/control_plane/pyproject.toml
[project]
name = "control-plane"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "docker>=7.0",
  "pydantic>=2.6",
  "sse-starlette>=2.1",
  "torch>=2.4.0",
  "numpy>=2.2.6",
  "pyyaml>=6.0.3",
  "httpx>=0.27",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "slow: long-running tests (loads the model)",
  "docker: tests that require a live docker daemon",
]
